moved the function to get project tasks for delete confirmation from tasks to projects
