+++ backend/projects/views.py
@@ -13,8 +13,9 @@ from django.forms.models import model_to_dict
 from django.db import IntegrityError
 
 import notifications
+from shoonya_backend.pagination import CustomPagination
 from .utils import ocr_word_count
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse
 from rest_framework import status, viewsets
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
@@ -1510,6 +1511,115 @@ class ProjectViewSet(viewsets.ModelViewSet):
             return Response(
                 {"message": "Project does not exist"},
                 status=status.HTTP_404_NOT_FOUND,
+            )
+
+    @swagger_auto_schema(
+        method="post",
+        request_body=openapi.Schema(
+            type=openapi.TYPE_OBJECT,
+            properties={
+                "project_task_start_id": openapi.Schema(type=openapi.TYPE_INTEGER),
+                "project_task_end_id": openapi.Schema(type=openapi.TYPE_INTEGER),
+                "project_task_ids": openapi.Schema(
+                    type=openapi.TYPE_ARRAY,
+                    items=openapi.Items(type=openapi.TYPE_INTEGER),
+                ),
+            },
+            description="Either pass the project_task_start_id and project_task_end_id or the project_task_ids in request body",
+        ),
+        manual_parameters=[
+            openapi.Parameter(
+                "id",
+                openapi.IN_PATH,
+                description=("A unique integer identifying the project"),
+                type=openapi.TYPE_INTEGER,
+                required=True,
+            )
+        ],
+        responses={
+            200: "returned successfully! or No rows to show",
+            403: "Not authorized!",
+            400: "Invalid parameters in the request body!",
+        },
+    )
+    @action(
+        detail=True,
+        methods=["POST"],
+        url_path="get_all_tasks_for_confirmation_before_deletion",
+        url_name="get_all_tasks_for_confirmation_before_deletion",
+    )
+    def get_all_tasks_for_confirmation_before_deletion(self, request, pk=None):
+        project = Project.objects.get(pk=pk)
+        try:
+            if not (
+                (
+                    request.user.role == User.ORGANIZATION_OWNER
+                    or request.user.is_superuser
+                )
+                and (request.user.organization == project.organization_id)
+            ):
+                return JsonResponse(
+                    {"message": "You are not authorized to access the endpoint."},
+                    status=status.HTTP_403_FORBIDDEN,
+                )
+
+            if "project_task_ids" in request.data:
+                project_task_ids = request.data.get("project_task_ids")
+                if len(project_task_ids) == 0:
+                    return JsonResponse(
+                        {"message": "No task Ids found in request"},
+                        status=status.HTTP_400_BAD_REQUEST,
+                    )
+            else:
+                project_task_start_id = request.data.get("project_task_start_id")
+                project_task_end_id = request.data.get("project_task_end_id")
+
+                if (
+                    project_task_start_id == ""
+                    or project_task_end_id == ""
+                    or project_task_start_id == None
+                    or project_task_end_id == None
+                ):
+                    return JsonResponse(
+                        {"message": "Project task start id or end id not found."},
+                        status=status.HTTP_400_BAD_REQUEST,
+                    )
+
+                project_task_ids = [
+                    id for id in range(project_task_start_id, project_task_end_id + 1)
+                ]
+
+            project_tasks = (
+                Task.objects.filter(project_id=project)
+                .filter(id__in=project_task_ids)
+                .order_by("id")
+            )
+
+            # pagination
+            paginator = CustomPagination()
+            result_page = paginator.paginate_queryset(project_tasks, request)
+
+            num_project_tasks = len(project_tasks)
+
+            if num_project_tasks == 0:
+                return JsonResponse(
+                    {
+                        "message": "No rows to show corresponding to the entered task ids"
+                    },
+                    status=status.HTTP_200_OK,
+                )
+            else:
+                serializer = TaskSerializer(result_page, many=True)
+                return paginator.get_paginated_response(
+                    {
+                        "message": "Project tasks retrieved successfully for valid task ids",
+                        "data": serializer.data,
+                    }
+                )
+
+        except Exception as error:
+            return JsonResponse(
+                {"message": str(error)}, status=status.HTTP_400_BAD_REQUEST
             )
 
     @is_project_editor
